Move looksLikeMastodonUrl to LinkHelper
